+++ Changelog.md
@@ -5,6 +5,7 @@
 - Add 'mindless' defense trait
 - Add 'NONE' option to traits field of Active Effects, to ignore AE if the attacker/defender does have one or more of the lists traits.
 - Fix name in "exchange cards" dialog.
+- Add exploding 'd20' and 'Bonus Die' roll buttons to the Torg control buttons.
 
 ## 13.29.0 - Card Hands Manager window